backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pymavlink==2.4.37
numpy==1.24.3
python-dotenv==1.0.1 